backend/requirements.txt
redis==4.5.1  # https://github.com/redis/redis-py
hiredis==2.2.1  # https://github.com/redis/hiredis-py
loguru==0.6.0
pytest==7.2.2
openai==0.27.2

# Django
# ------------------------------------------------------------------------------
django==4.1.7  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-ninja==0.21.0  # https://github.com/vitalik/django-ninja
pytest-django==4.5.2

# Django REST Framework
django-cors-headers==3.13.0  # https://github.com/adamchainz/django-cors-headers


# Deployment
gunicorn==20.1.0

